% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_scalar.R
\name{is.scalar}
\alias{is.scalar}
\title{Determine if object is scalar.}
\usage{
is.scalar(x)
}
\arguments{
\item{x}{a vector of length 1}
}
\value{
A logical TRUE or FALSE
}
\description{
Simple checker if the value is 1-dimensional
}
\examples{
is.scalar(1)
is.scalar(c(1,2))
}
