% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_streamlined.R
\name{optimize_designs}
\alias{optimize_designs}
\title{Streamlined Design Optimization Pipeline
Authors: Josh Betz (jbetz@jhu.edu) and Michael Rosenblum}
\usage{
optimize_designs(ui.n.arms, ui.type.of.outcome.data,
  ui.time.to.event.trial.type, ui.time.to.event.non.inferiority.trial.margin,
  ui.subpopulation.1.size, ui.total.alpha, ui.max.size, ui.max.duration,
  ui.accrual.yearly.rate, ui.followup.length, ui.optimization.target,
  ui.time.to.event.censoring.rate, ui.mcid, ui.incorporate.precision.gain,
  ui.relative.efficiency, ui.max.stages, ui.sann,
  ui.include.designs.start.subpop.1, ui.population.parameters, ui.desired.power,
  ui.scenario.weights, default.pct.of.max = 0.8, min.n.per.arm = 25,
  min.enrollment.period = 0.5, default.function.scale = 1,
  default.n.scale = 100, default.period.scale = 2,
  default.max.iterations = 1000, default.n.simulations = 10000,
  default.means.temperature = 100, default.survival.temperature = 10,
  default.evals.per.temp = 10, default.report.iteration = 1,
  default.power.penalty = 1e+05, default.boundary.to.enroll = 1)
}
\arguments{
\item{ui.n.arms}{Number of Arms (including control arm), e.g., 2 arms means one treatment arm and one control arm}

\item{ui.type.of.outcome.data}{"continuous", "binary", or "survival"}
}
\value{
List of optimized designs
}
\description{
Streamlined Design Optimization Pipeline
Authors: Josh Betz (jbetz@jhu.edu) and Michael Rosenblum
}
\examples{
optimize_designs(
  ui.n.arms=2,
  ui.type.of.outcome.data="time-to-event",
  ui.time.to.event.trial.type="non-inferiority",
  ui.time.to.event.non.inferiority.trial.margin=1.35,
  ui.subpopulation.1.size=0.33,
  ui.total.alpha=0.05,
  ui.max.size=10000,
  ui.max.duration=10,
  ui.accrual.yearly.rate=1000,
  ui.followup.length=0,
  ui.optimization.target="size",
  ui.time.to.event.censoring.rate=0,
  ui.mcid=0.1,
  ui.incorporate.precision.gain=FALSE,
  ui.relative.efficiency=1,
  ui.max.stages=2,
  ui.sann=10,
  ui.include.designs.start.subpop.1=FALSE,
  ui.population.parameters= 0.08*matrix(c(1.00, 1.00, 1.00, 1.00, 1.00, 1.00, 1.00, 1.3500001, 1.00, 1.00, 1.00, 2.14, 1.00, 1.00, 1.3500001, 1.3500001), ncol=4, byrow=TRUE,,dimnames=list(c(),c("lambda1_con","lambda2_con","lambda1_trt","lambda2_trt"))),
  ui.desired.power=0.8*matrix(c(1.00, 1.00, 0, 1.00, 0, 0, 1.00, 0, 0, 0, 0, 0), ncol=3, byrow=TRUE,dimnames=list(c(),c("Pow_H(0,1)","Pow_H(0,2)","Pow_Reject_H0,1_and_H0,2"))),
  ui.scenario.weights=matrix(rep(0.25,4),ncol=1,dimnames=list(c(),c("weight"))),
    default.max.iterations=2,
  )
}
