% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_temperature.R
\name{sa.temperature}
\alias{sa.temperature}
\title{Tempature of Cooling Schedule}
\usage{
sa.temperature(iteration = 1, max.iterations = 10000, temperature = 10,
  evals.per.temp = 10)
}
\arguments{
\item{evals.per.temp}{}
}
\description{
Thhis is a function for viewing the default cooling schedule for
}
\examples{
sa.temperature(iteration=20, max.iterations=1e4,
temperature=10, evals.per.temp=10)
plot(sa.temperature(iteration=seq(1, 1e4), max.iterations=1e4,
temperature=10, evals.per.temp=500)~seq(1, 1e4), type='l',
ylab="SA Temperature", xlab="Iteration")
}
